Fix wrong unique transformation on command list.
When using unique we were creating a unique list of command classes, not command names from the config file.

We were not able to give use a class multiple times with different names (like alias). This tries to fix this issue.

Also currently only detected if name of command was completely left off. We also need to check in case the command name was left empty like ''. This is not picked up using the unique technique.