Fix nasty bug.
If CommandHandler is instantiated via the container, it gets build using the DEFAULT bot config information.

If the bot responding to the command is NOT the default bot then none of the correct commands etc get loaded.

Therefore we have to build the CommandHandler using the correct Bot Config information.