Add mixin to EventFactory
WIth static analysis enabled on the project, and the latest changes to the laravel package `WebhookController`, we now call a method `hasListeners` on this event factory. This event factory uses magic `__call` to pass undefined methods to the underlining Illuminate Dispatcher class.

Without the mixin tag the static analysis does not know we are doing this and so thinks we are calling an undefined method. This fixes this issue.